Fix some invalid left shifts (undefined behavior)
In src/audio.c, samples may be negative and left shifting a negative
value is undefined behavior (signed overflow). Unsigned left-shift is
well-defined and produces the desired results, so use an unsigned
operand.

In src/video.c the left-shift right-side operands are negative in some
cases, the results are which are undefined. In practice, on x86 hosts it
typically behaves as though masked per this commit. The mask preserves
the most likely current behavior using well-defined semantics.

There is one more frequent left-shift signed overflow in ymfm_opm.cpp
(opm_registers::opm_registers, shifting pm). Since this is a vendored
dependency I haven not included a fix in this commit, but fixing it is
similarly simple.

All caught with Undefined Behavior Sanitizer (-fsanitize=undefined).