Various minor header fixups (#300)
* src/files.c: stdio.h for printf
* src/memory.c: inttypes.h for PRIu64
* src/sdcard.c: string.h for strncpy
* src/testbench.c: string.h for strncpy
* src/utf8_encode.h: include guard
* src/video_win32.c: de facto standard case for windows.h
* src/wav_recorder.c: stdio.h for printf

I ran into all but one of the above compiling with Mingw-w64. The case
for windows.h is important for cross-compilation, and on case-sensitive
file systems it's always all lowercase. The missing include guard in
utf8_encode.h isn't normally a problem, but it affects my custom unity
build and its omission is likely an accident.