via2: clean up CA1/CB1 interrupt handling
in particular, if the user perhipheral connected pins don't include
CA1 or CB1, don't check them for level changes. Old logic would treat
unconnected as 'low' which could, in principle, override a value set
by other means. In practice, there is no code path for those values to
be set elsewhere, but no sense in leaving in a future bug.